+++ lib/td_dd/canada/data_structure_type_abilities.ex
@@ -1,0 +1,13 @@
+defmodule TdDd.Canada.DataStructureTypeAbilities do
+  @moduledoc false
+  alias TdDd.Accounts.User
+  alias TdDd.DataStructures.DataStructureType
+
+  def can?(%User{}, :index , %DataStructureType{}), do: true
+
+  def can?(%User{}, :index , DataStructureType), do: true
+
+  def can?(%User{}, _action, %DataStructureType{}), do: false
+
+  def can?(%User{}, _action, DataStructureType), do: false
+end