+++ lib/td_dd_web/controllers/data_structure_type_controller.ex
@@ -1,0 +1,146 @@
+defmodule TdDdWeb.DataStructureTypeController do
+  use TdDdWeb, :controller
+  use PhoenixSwagger
+
+  import Canada, only: [can?: 2]
+
+  alias TdDd.DataStructures.DataStructuresTypes
+  alias TdDd.DataStructures.DataStructureType
+  alias TdDdWeb.SwaggerDefinitions
+
+  action_fallback(TdDdWeb.FallbackController)
+
+  def swagger_definitions do
+    SwaggerDefinitions.data_structure_type_definitions()
+  end
+
+  swagger_path :index do
+    description("Get sources of the given type")
+    produces("application/json")
+
+    response(200, "OK", Schema.ref(:DataStructureTypesResponse))
+    response(403, "Forbidden")
+    response(422, "Client Error")
+  end
+
+  def index(conn, _params) do
+    user = conn.assigns[:current_user]
+
+    if can?(user, index(%DataStructureType{})) do
+      data_structure_types =
+        DataStructuresTypes.list_data_structure_types()
+        |> Enum.map(&DataStructuresTypes.enrich_template/1)
+
+      render(conn, "index.json", data_structure_types: data_structure_types)
+    else
+      conn
+      |> put_status(:forbidden)
+      |> put_view(ErrorView)
+      |> render("403.json")
+    end
+  end
+
+  swagger_path :create do
+    description("Creates a new data structure type")
+    produces("application/json")
+
+    parameters do
+      data_structure_type(:body, Schema.ref(:CreateDataStructureType), "Parameters used to create a data structure type")
+    end
+
+    response(200, "OK", Schema.ref(:DataStructureTypeResponse))
+    response(403, "Forbidden")
+    response(422, "Client Error")
+  end
+
+  def create(conn, %{"data_structure_type" => data_structure_type_params}) do
+    user = conn.assigns[:current_user]
+
+    with true <- can?(user, create(%DataStructureType{})),
+         {:ok, %DataStructureType{} = data_structure_type} <-
+           DataStructuresTypes.create_data_structure_type(data_structure_type_params) do
+      conn
+      |> put_status(:created)
+      |> render("show.json", data_structure_type: data_structure_type)
+    end
+  end
+
+  swagger_path :show do
+    description("Get data structure type with the given id")
+    produces("application/json")
+
+    parameters do
+      id(:path, :string, "id of Data Structure Type", required: true)
+    end
+
+    response(200, "OK", Schema.ref(:DataStructureTypeResponse))
+    response(403, "Forbidden")
+    response(422, "Client Error")
+  end
+
+  def show(conn, %{"id" => id}) do
+    user = conn.assigns[:current_user]
+
+    with true <- can?(user, show(%DataStructureType{})),
+         data_structure_type <- DataStructuresTypes.get_data_structure_type!(id) do
+      render(conn, "show.json", data_structure_type: data_structure_type)
+    else
+      false ->
+        conn
+        |> put_status(:forbidden)
+        |> put_view(ErrorView)
+        |> render("403.json")
+    end
+  end
+
+  swagger_path :update do
+    description("Updates Data Structure Type")
+    produces("application/json")
+
+    parameters do
+      id(:path, :string, "id of Data Structure Type", required: true)
+      data_structure_type(:body, Schema.ref(:UpdateDataStructureType), "Parameters used to update a Data Structure type")
+    end
+
+    response(200, "OK", Schema.ref(:DataStructureTypeResponse))
+    response(403, "Forbidden")
+    response(422, "Client Error")
+  end
+
+  def update(conn, %{"id" => id, "data_structure_type" => data_structure_type_params}) do
+    user = conn.assigns[:current_user]
+
+    with true <- can?(user, update(%DataStructureType{})),
+         data_structure_type <- DataStructuresTypes.get_data_structure_type!(id),
+         {:ok, %DataStructureType{} = data_structure_type} <-
+           DataStructuresTypes.update_data_structure_type(
+             data_structure_type,
+             data_structure_type_params
+           ) do
+      render(conn, "show.json", data_structure_type: data_structure_type)
+    end
+  end
+
+  swagger_path :delete do
+    description("Deletes a Data Structure Type")
+
+    parameters do
+      id(:path, :string, "Data Structure id", required: true)
+    end
+
+    response(204, "No Content")
+    response(403, "Forbidden")
+    response(422, "Client Error")
+  end
+
+  def delete(conn, %{"id" => id}) do
+    user = conn.assigns[:current_user]
+
+    with true <- can?(user, delete(%DataStructureType{})),
+         data_structure_type <- DataStructuresTypes.get_data_structure_type!(id),
+         {:ok, %DataStructureType{}} <-
+           DataStructuresTypes.delete_data_structure_type(data_structure_type) do
+      send_resp(conn, :no_content, "")
+    end
+  end
+end