+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - [TD-911] Allow to limit lineage/impact levels 
 
+### Fixed
+
+- [TD-2826] DataStructures.list_data_structures can receive domain_id list
+
 ## [4.0.0] 2020-07-01
 
 ### Changed