+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-911] Allow to limit lineage/impact levels 
+- [TD-2322] Allow to search all data structures versions without 10_000 limit using ES scroll API
+
 ### Fixed
 
 - [TD-2826] DataStructures.list_data_structures can receive domain_id list